this is one of the output files of googlemaps_plots.py
This is plot of Divvy stations. 
Divvy stations are plotted in either red or blue.
If number of user that rented bikes from a location is higher than average, it's plotted in red.
If number of user that rented bikes from a location is lower than average, it's plotted in blue.